Bug 1539021 - Specfify all packages during upgrade
When we update package "a-${version}" and a requires b >= ${version} if we
don't specify the version of b yum will choose the latest version of b
available and the whole set of dependencies end up at the latest version.
Since the package module, unlike the yum module, doesn't flatten a list
of packages into one transaction we need to do that explicitly. The ansible
core team tells us not to rely on yum module transaction flattening anyway.