Increase watch_retry_timeout for kuryr-daemon
When deploying on OpenStack, we're putting OpenShift API behind Octavia
load balancer. Octavia is sometimes very slow with allowing the traffic
to pass.

Now kuryr-daemon container is connecting to OpenShift API through that
load balancer. In case of LB not being ready at the moment it retries
for 60 seconds and then stops. For Octavia this is sometimes not enough.

This commit increases that timeout to 3600 seconds, to make kuryr-daemon
retry virtually forever.