Fix openshift_logging on Python3
Popen returns a byte object on Python 3 so it will fail when being
compared to a string. Since we ever only use it as a string we simply
cast it right after creation.

When converting to a string I decided to use the `replace` error handler
which will replace invalid UTF bytes with `?` instead of raising an
exception. Another option would be to use `ignore` but that could cause
confusion further down as the string would not really be a proper UTF
string.

Closes #4163