Assign openshift_version from openshift_pkg_version with correct format
On release-3.6 `openshift_version` is assigned by
`openshift_pkg_version` when openshift_version was not set but
openshift_pkg_version was. However, The format of
`openshift_pkg_version` is different from openshift_version and could
contain `-` and `*` infront of the value. Due to this, some tasks
using openshift_version fail unexpectedly (e.g - bz#1592111).

To fix this issue (without breaking current behavior), this patch adds
normalization of openshift_version only when it was assigned by
openshift_pkg_version as:

  |openshift_pkg_version            | openshift_version  |
  |---------------------------------|--------------------|
  | -3.6.173.0.117-1.git.0.b1f7e54  | 3.6.173.0.117      |
  | -3.6.173.0.117                  | 3.6.173.0.117      |
  | -3.6*                           | 3.6                |