Increate lbaas_activation_timeout for kuryr-controller
When using Octavia as LBaaS on slow (or busy) environments it may take
longer than the default 300 seconds for the Amphora VM to be
provisioned.

If this timeout is reached, kuryr will retry the action. However, as the
LBaaS creation was already triggered and due to a bug in Octavia
(https://storyboard.openstack.org/#!/story/2001944) when trying to list
existing load balancers with filtering, the kuryr controller will fail
to perform the remaining operations.

This commit increases that timeout to 1200 seconds by default.