Provision OpenStack node networking with Ansible
This adds playbooks for provisioning the node/control plane network
using the Ansible OpenStack modules (as opposed to doing it in Heat).
Only the resources required for OpenShift SDN are being created right
now. The idea is to get to a working OpenShift as fast as we can and add
the extra stuff such as Kuryr support later.

It also adds stub tasks for the extra resources.