Support oreg_url for default image prefix of webconsole
Even though webconsole is a core component and it has a prefix of
`registry.access.redhat.com/openshift3/ose-*` or
`docker.io/openshift/origin-`, it does not take oreg_url for the
prefix.

This patch adds oreg_url for openshift_web_console_prefix, but does
not make any change for default value.