Install metrics-server pre-upgrade, if metrics
This installs metrics-server pre-upgrade, so that we don't break
existing HPAs.  In both the case of fresh installation and upgrade, we
deploy metrics-server if the user has configured metrics to be deployed,
so that we don't break existing installations.  The user can still
manually set `openshift_metrics_server_install=False` to disable this
behavior.