Update docker registry auth to idempotent
Currently, the primary method for placing registry
authentication credentials for use with oreg_url
is via a command call to 'docker login'

This cli call is not very idempotent.

This commit changes primary method to custom module
that creates / updates the json file directly in
an idempotent fashion and utilizes urllib to test
that the credentials are valid before proceeding.