Remove alternative oreg vars and update logic
This commit removes oreg_url_master and oreg_url_node
in favor of only retaining oreg_url.  It is still possible
to use inventory to have a different value for oreg_url
for masters and nodes, if desired.

Additionally, this commit fixes logic for oreg_host,
which should account for new enterprise registry.