Fix order for invoking the hostpath storage task for registry
When attempting to using the hostpath storage for the registry for
an all-in-one setup, the playbook would fail due to an ordering issue
that caused it to not be able to correctly wait for the pods to come
online after configuring.

This change moves it to the correct place, and unbreaks hostpath storage
for all-in-one PoC configurations.