Add basic support for a separate baremetal compute role via OpenStack. fixes #9399
If we want to be able to have a mixture of virtual computes and physical computes in a
cluster built on OpenStack we need to be able to specify different flavors:
- a flavor for the virtual compute
- a flavor for the baremetal (physical server) compute

This change takes a first cut at supporting both virtual and physical compute nodes in the
same cluster.  The work is on the OpenStack side as all the OpenShift installer needs is
for the nodes to be added to the Ansible inventory.

3 new variables were added to support this feature:
openshift_openstack_num_baremetal_nodes: 1
openshift_openstack_baremetal_node_hostname: baremetal-node
openshift_openstack_baremetal_node_flavor: baremetal

Known Limitations:
In this configuration, cinder volumes are not available to the baremetal servers.
Devising logic to create the volume and mount it on virtual compute nodes but
not on physical compute nodes was challenging.  In the end I opted to create a separate
heat_stack_baremetal_server.yaml.  This was more straightfoward and may prove to
helpful for other items like networking in the future.

Today this is working and testing with provider networks only.

There is support for naming the baremetal servers differently for easier identification, however
no additional tagging is done at this point to provide info for OpenShift scheduling.