Refactor registry-console template and vars
This commit refactors registry-console role to
create it's own template, updates the template to
utilize proper image name string, and conforms
component to utilizing oreg_url.

This commit also removes redundant openshift_hosted_templates
role as it is no longer needed with this commit.

This commit also ensures nodeSelector is set to master.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1608476
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1613647