Remove the reconcile-sccs check
The reconcile-sccs command is being called with the `oc` binary
of the version to upgrade to. This does not make sense because
such a check would generate false positive every time the bootstrapped
SCCs change in the newer version (e.g. add a non-default value
to a new field).

Related to: https://bugzilla.redhat.com/show_bug.cgi?id=1610496