Remove atomic-openshift-master package during upgrade
During an upgrade from 3.9, when the node packages are updated, the
atomic-openshift package has a dependency on atomic-openshift-master
package.  This results in the master package being updated to 3.10.
This package is not necessary on 3.10 and is removed during the control
plane upgrade.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1572857