GlusterFS: Fix heketi_pod check
The previous solution for tracking which heketi pods are active was broken
with the merging of bb5888854. In a two-cluster deployment of GlusterFS
(`glusterfs` and `glusterfs_registry` groups defined) the second cluster
would erroneously receive information for which heketi pods were running from
the first deployment, due to problems with how set_fact was used. This commit
overhauls that solution to fix the problem as well as a few more that arose in
development.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1623777
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1626751

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>