check for sync pod availability before node selector
The sync pods need to be running on all nodes to ensure that labels are
up to date before checking for nodes matching a selector.  This change
moves the location of the sync pod check, because previously the sync
check would succeed before infra and compute nodes are added to the cluster.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1609019