Fix kuryr support for custom OpenStack network and subnet
The addition of [1] didn't fully account for the changes needed when
deploying with kuryr. This PR ensures that kuryr can be deployed without
FIPS too.

In addition, it moves from requiring the custom network and subnet names used
to require the created subnet and router. Note the network can be obtain
from the subnet, and the router is needed for kuryr support.

[1] https://github.com/openshift/openshift-ansible/commit/22b6b19941517e8cae584e98ca7edc00aa426c3d#diff-aaccff40c391dc9989e2ca0acf56aca4