Retry our etcd health check
This is a followup attempt to fix a bug. Originally dns was failing due
to problems with the sdn pod. Now it appears that etcd is just prone to
taking longer than expected to become healthy. Wait up to 3 minutes for
etcd to come back online

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1623145