Reload tuned service when node-config.yaml has changed.
One of the tasks of the openshift-ansible installer is to include the tuned
role and apply the correct per-node tuning.  The tuned role relies on the
"recommend" tuned daemon functionality, which queries node-config.yaml for
appropriate node labels and sets the correct tuned profile according to these
labels.  Unfortunately, with the introduction of the openshift-node/sync pods,
the node labelling information may be set by the sync pods after the installer
finished and therefore OpenShift nodes may run with incorrect tuned profiles.

This PR adds a functionality to restart the tuned daemon and as a result set
the correct tuned profile when node-config.yaml changes.

Fixes:
- https://bugzilla.redhat.com/show_bug.cgi?id=1569917
- https://bugzilla.redhat.com/show_bug.cgi?id=1612776