Make cloud-user SSH key maintenance more reliable
Check GCP project metadata for cloud-user's SSH key more specifically. Before,
the key for cloud-user would only be updated if the key's contents were present
anywhere in the metadata value; sometimes, the key was present for the 'default'
user, but missing from cloud-user, causing the script to skip re-adding the key
for cloud-user, making the instances inaccessible to ansible.