Ensure master facts are set during node scale-up
If the cached facts file (openshift.facts) on masters is missing or
does not contain the 'master' section, the necessary master facts will
not be created during the fact initialization in std_include.yml.

This commit adds a step to ensure all master facts are set prior to
performing node scaleup.

Bug: 1627736