Start only the ovsdb so we can add the config safely
The existing code starts both the ovsdb and the ovsvswitchd and then
chances the database to limit the threads used.

Unfortunately, on very fast machines that have many cores, there is a
race between when the config change gets in to limit the threads used,
and when the threads get started.  If the config change loses the
race, then the threads aren't limited and the vswitchd becomes
unresponsive.

The fix is to start the db component, program it, then start the
vswitchd component.

Fixes bug 1633892 (https://bugzilla.redhat.com/show_bug.cgi?id=1633892)