Allow wiping GlusterFS devices at install time
Extract glusterfs wipe into its own playbook and call it from the
install playbook as well.

This allows to redeploy OpenShift on the same nodes without the need to
call the uninstall playbook prior to it.

This is also more in line with the documentation  for the
`openshift_storage_glusterfs_wipe` variable that didn't mention wiping
GlusterFS storage devices only happened when specifically calling the
uninstall playbook.