Only upgrade ASB and TSB when they are installed
There are dedicated inventory vars to control whether SC, ASB or TSB
should be deployed. However during the upgrade process, for all 3
components the variable for the SC is queried.

This has 2 undesired side effects for setups where ASB and TSB are not
deployed:
1. The ASB is always deployed during upgrade, no matter what
2. The TSB upgrade role fails because at one point it waits for the TSB
(which is not deployed) to become ready