Removing run_once usage on include_tasks
Previously, run_once (and several other options) were silently ignored on include_tasks
and effectively did nothing. As of 2.7/2.8 these invalid options will now raise errors,
so they must be removed.

This was tested against an install of 3.11 originally, so there may be others that were
added after that release.