Set maxSurge in kuryr-controller Deployment
As kuryr-controller normally doesn't support HA, we should make sure
there can be only one instance of that service in the deployment.
Otherwise we might end up with orphaned OpenStack resources left as two
controllers fight over the events.

This commit sets maxSurge=0 and maxUnavailable=1 in kuryr-controller's
Deployment rollingUpdate strategy. This way we prevent a situation when
on Deployment update (e.g. when changing the kuryr-controller image) a
second instance of kuryr-controller gets created before the old one is
terminated.