Extend Openstack cloud provider LB parameters
This commit enables control of the following
parameters in  Openstack cloud provider Loadbalancer:

Use Octavia:  which OpenStack LBaaS endpoint
(Octavia or Neutron LBaaS) should be used.

Floating IP network: the network ID in which floating IPs
will be allocated from. If not specified, the first network
that marked as external will be used.