openshift_ovirt: add vms into inventory groups
Motivation:
This role should be easy to consume and should be ready made to be
consumed with other playbooks to install a cluster without specifing an
explicit inventory of openshift nodes. The manifest should be the single
point where one specifies the nodes layout.

Modification:
2 new tasks added to create the inventory from the manifest -
create_inventory_from_vms.yaml and
create_all_in_one_inventory_from_vms.yaml. The latter will take the
first master and make it all-in-one, regardless of the count.

Result:
The role now supports all-in-one install by setting
openshift_ovirt_all_in_one=true.
To install the a cluster, just include the role, and the prerequisit and
install_cluster playbooks - see the examples in the
roles/openshift_ovirt/README.md

Signed-off-by: Roy Golan <rgolan@redhat.com>