glusterfs: bind-mount /dev/mapper into the glusterfs-server container
Commit f9445033 removed the bind-mount for /dev as this will be setup by
the container runtime automatically. However it seems that /dev/mapper
is missing, prevending users to supply /dev/mapper/* device names to
Heketi. It also seems possible that not all(?) LVM devices are available
when a container starts.

This change adds /dev/mapper as an additional bind-mount. CRI-O does
prevent /dev to be bind-mounted, but subdirectories of /dev do not seem
to be an issue.

Updates: gluster/gluster-kubernetes#499