Install node package names based on openshift_release
Previously, the packages to be installed were hard coded for origin. As
a result, there was no way of installing the openshift-enterprise
packages. This was a regression from 3.x behavior.

This patch uses the openshift_service_type variable to determine which
packages are to be installed based on the value of openshift_release.