Add openshift_bootstrap_endpoint for scaleup.
Prior to this change, the install-config.yaml needed to be retained from
the original openshift-install of the cluster and passed to the scaleup
playbook via the openshift_install_config_path variable.

This change allows scaleup to work without the install-config.yaml by
passing the openshift_bootstrap_endpoint variable directly as part of the
inventory while maintaining backwards compatability.