gluster: detect intent to deploy legacy OpenShift Container Storage
Old versions of OpenShift Container Storage (OCS) do not have the needed
scripts to handle the HOST_DEV_DIR environmant variable. These versions
should (try to) use a normal /dev bind mount from the host into the
container.

This will not work in combination with CRI-O. Only OCS-3.11.1 and newer
have the HOST_DEV_DIR script.

Upstream Gluster containers do not have this problem, as the images have
been updated to contain the script. Hence there is a check for the
"openshift-enterprise" variant of deploying.