Configure API aggregation in serial
Previously, the API aggregation changes made during an upgrade were
performed in parallel. This caused unexpected service impact when the
entire control-plane was started at the same time.

This change forces the API aggregation to happen in series of 1. As a
result, clusters with a multiple node controle-plane will not experience
the same service impact.