Bug 1671315 - Kibana and Curator pods names are with ops
I basically changed all of the logging rules to use include_role
instead of import_role.  In my (albeit limited) testing of
the logging roles with ansible 2.6 and 2.7, logging seems to
work just fine, and it fixes all of the "-ops" naming
problems.  This allows the logging roles to work with both
ansible 2.6 and ansible 2.7.

The problem is that the import_role behavior changed between
ansible 2.6 and ansible 2.7:
https://docs.ansible.com/ansible/latest/porting_guides/porting_guide_2.7.html?highlight=import_role#include-role-and-import-role-variable-exposure
import_role is changing global variables like
openshift_logging_kibana_hostname when the role is imported,
even if the role is conditionally imported and the condition
is false.