tmpfile.d file clears the cni config on boot
Networking is ready when the /etc/cni/net.d/80-openshift-network.conf
exists. When the host is rebooted or power cycled, this file is
still on the disk and the cluster immediately starts creating pods.
This change deletes the file before the cluster node is started.

roles/openshift_node/templates/cleanup-cni.j2 is the tmpfile.d file that does the delete,
tasks/systemd_units.yml is the playbook that references it.

Add a lifecycle: preStop hook to the delete the files.

Also deletes ovs db on boot (contents are invalid after boot)

SDN-329 - Create a systemd unit that clears the cni configuration directory on boot
https://jira.coreos.com/browse/SDN-329

Part of solution to:
Bug 1654044 - OCP 3.11: pods end up in CrashLoopBackOff state after a rolling reboot of the node

Signed-off-by: Phil Cameron <pcameron@redhat.com>