openshift_ovirt: Fix hostname in the inventory groups
Fixing the hostname passed to add_host. The bug was that in the case of
multiple hosts a pattern was used, like in an inventoy file:

  [masters]
  master[0:2].local

While this works in an inventory file, patterns are not supported
in the in-memory inventory and this simply fails with unreachable
hostname(master[0:2].local will not resolve).

Instead, we have to loop in the add_host module and insert every
host with the groups details etc.

Signed-off-by: Roy Golan <rgolan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1698922