Avoid duplicate env vars in router deploymentconfig
Prevent edits from causing duplicate environment variables in the
deploymentconfig of a prepared router.

This commit fixes bug 1656487.

https://bugzilla.redhat.com/show_bug.cgi?id=1656487

* roles/lib_openshift/library/oc_adm_router.py (Router.add_modifications):
* roles/lib_openshift/src/class/oc_adm_router.py (Router.add_modifications):
If an edit appends an environment variable and a variable by that name
already exists in the deploymentconfig, delete the old variable.
* roles/lib_openshift/src/test/unit/test_oc_adm_router.py
(RouterTest.test_append_to_env): New unit test.  Verify that the
deploymentconfig of a prepared router that results from a router config
with edits that append environment variables has exactly 1 instance of each
appended variable, irrespective of whether the variable already exists in
the default deploymentconfig or whether more than 1 edit appends it.