Enable ports population for new namespaces
When using the namespace subnet driver, kuryr creates a new Neutron
subnet for each new namespace. This will require a new pool for the
pods allocated in that subnet, and therefore the pre-population of
ports done during the deployment is not useful. By enabling the new
kuryr kuryrnet handler, for each newly created namespace the handler
will create a new pool in each node so that pods deployment on that
new namespace are later speeded up.