Mount the parent dir in daemonset configuration for openshift-sdn
In the daemonset configuration for openshift-sdn the hostpath mount
specified is a file name instead of the parent dir. Kubernetes will
create the file but it will create it as a directory during pod setup.

This PR changes the hostpath mount to the parent dir and if a directory
by the name of origin-node exists, deletes it as a precaution.