Modify mounts when using atomic during imageconfig if oreg username is defined
When upgrading from 3.10 to 3.11 on if the registry is changed
to one that requires authentication the upgrade will fail
unable to pull images.  The cause is
there is no mount to `/root/.docker/` to the atomic-openshift-node
container.  This PR adds to the `playbooks/openshift-node/imageconfig.yml`
playbook an conditional to run a new playbook
`openshift-node/private/configure_system_container.yml`
if using atomic and oreg_auth_user is used.
The new playbook:
- grabs the current registry image url
- uses the existing openshift-node role task to update
the mount with the existing image.