Bug 1752532: upgrade from 3.11 failed to upgrade es settings from configmap
https://bugzilla.redhat.com/show_bug.cgi?id=1752532#c24
The problem is that the code that compares diffs of the configmap does
not know how to handle configmaps that are too old.  In that case,
and if the user has not manually modified the configmap, this introduces
a new option:

`openshift_logging_elasticsearch_replace_configmap`

By default this is `false`. If you set it to true, running ansible
will replace your `logging-elasticsearch` configmap with the current
default values.  It will also copy the current configmap to
`logging-elasticsearch.old` in case there are customizations you would
like to copy from the old one and apply to the new one.

Once you do this, you can remove `openshift_logging_elasticsearch_replace_configmap`
and let ansible use the diff and merge procedure, since the configmap
will now be up-to-date, and the diff and merge code will work.