Fix upgrade not respecting openshift_pkg_version
During an upgrade with openshift_pkg_version specified the latest
available version was being installed.  The openshift-clients package
was being upgraded during an unrelated master upgrade step.  Due to
package dependencies, this resulted in all dependent openshift packages
being updated to the latest version.  This change removes the
openshift_cli role as the clients are installed and upgraded as part of
the openshift_node role.  The upgrade of the client binaries on Atomic
Host were moved to the openshift_node role for consistency.