Extend docker proxy evaluation.
In some special situations it can be useful to be able to define an
extra proxy for the docker daemon.

Add possibility to use a different proxy for docker. If there is no
extra proxy for docker defined `openshift_http_proxy` and
`openshift.common.https_proxy` will be used as fallback.