Back up certificates on all masters
When redeploying certificates on masters, the generated-configs/ and
master/ directories in /etc/origin would be backed up only on the first
master.  Then the certs and configs in /etc/origin/master on all masters
would be removed.  This caused issues if the certificate redeploy
playbook failed and was later run again because certificates would be
missing.  To recover from this it required restoring the master
certificates which were located in the generated-configs backup on the
first master but it did not also include all the configs.

https://bugzilla.redhat.com/show_bug.cgi?id=1762932

This change separates the generated-configs/ backup from the master/
backup and runs the master/ backup on all masters to make the recovery
of removed certificates and configs easier. The backup filename has been
changed as follows:

Original name (first master only):
/etc/origin/master-node-cert-config-backup-{{ ansible_date_time.epoch }}.tgz

New generated-config backup name (first master only):
/etc/origin/generated-configs-backup-{{ ansible_date_time.epoch }}.tgz

New master certificate and config backup name:
/etc/origin/master-cert-config-backup-{{ ansible_date_time.epoch }}.tgz

(This name conforms to the same backup file created for the node cert and config.)