Fix htpasswd for Hawkular when openshift_metrics_hawkular_fips is set.
When setting openshift_metrics_hawkular_fips to true, the output of the
htwpasswd content would cause a failure due to trying to write to a
local tmp directory that doesn't exist on the remote host.

This fix updates the logic to output the content to the same location as
it would if openshift_metrics_hawkular_fips were set to false.

Signed-off-by: Jared Hocutt <jhocutt@redhat.com>