possibility to patch configmaps on remote
Cause of a missing patch command on my ansible controller wich is
running in a container, I needed the possibility to patch the configmaps
on the remote host which got the patch package. So I've added this
ability with an additional variable called
"openshift_logging_patch_configmaps_on_remote" (Default is FALSE).