Fix test_image for additional registries
Fixes an issue creating docker credential files if both oreg_url
and openshift_additional_registry_credentials are provided.

Docker authentication credentials are created on the nodes by a custom
module, docker_creds.  The module is run for oreg_url if provided,
and openshift_additional_registry_credentials if provided.
If oreg_url is provided, credentials will be created using a test image

based on oreg_url.  If openshift_additional_registry_credentials is
provided, credentials will be created based on a default registry url,
or the test_image provided with the credentials.

When updating variables to use a default test image, additional
refactoring was done to improve the naming convention of variables only
used within the roles/playbooks.  No user-facing variables were changed.

For clarity of the source of variables used outside of openshift_facts,
the variables have been prefaced with "openshift_facts_".