gluster: configure pod template to enable support for nsenter
The newly added /usr/sbin/exec-on-host wrapper calls 'nsenter' to run
commands (intended for LVM) on the host, instead of in the container.
This will only work correctly in case the /-filesystem of the host is
mounted on $HOST_ROOTFS, which defaults to /rootfs, and the container
can access the PID-namespace of the host (to get to PID 1).

Signed-off-by: Niels de Vos <ndevos@redhat.com>