Remove cert expiry check during cert redeploy
When redeploying certificates, the cert expiry check provides little
value because the expectation is that the certificates will be replaced.
Addtionally, there are situations where certificates are in an invalid
state and redploy is blocked by the check.  Removing the checks will
allow certificate redeploy to proceed without requiring additional
inventory vars to override expiry days or invalid/missing certificates.