Create configmaps for all items in openshift_node_groups
In some circumstances such as provisioning in AWS with autoscaling
groups, the nodes may not exist in the inventory.  As such, the
configmaps for those nodes will not be created during install.  This
change reverts the code to the original behavior of creating configmaps
for all names listed in openshift_node_groups.  If it is desired that
the default configmaps are not created, openshift_node_groups should be
specified in the inventory to limit to only the configmaps desired.