Fix machine creation failure message
In CI we use machine sets to provision RHEL worker instances in AWS.
The rescue clause of the machine creation tasks should summarize the
failure state if one of these tasks fails.  The failure message was
referencing the 'resources' object which was previously used when using
the k8s Ansible modules and therefore was not valid because the creation
tasks have been moved to 'command' which provides 'stdout'.  The
status.errorMessage object is not present during some failures, so the
error detail has been removed.  Previous task output can be used to
determine the cause of the failure.