Detect an incomplete OpenShift CA redeployment
Adds a check of the master-config.yaml to determine if the client.CA has
been reverted.  If not, the play will fail indicating
openshift_redeploy_openshift_ca=true must be set in the inventory.

This check will prevent inadvertant certificate redeploy when the
OpenShift CA has been updated and not rolled out.