openshift-cluster/upgrades: Allow short nodes names when filtering
When filtering the list of nodes to upgrade based on using the
openshift_upgrade_nodes_label variable, the node names may not match if
the kubelet returns the short node name.

 - `hostname` = fqdn          (infra-0.example.com)
 - `hostname -f` = fqdn       (infra-0.example.com)
 - `hostnamectl` = fqdn       (infra-0.example.com)
 - `oc get node` = short name (infra-0)

This changes the filter to test if the short host name is in the node
name returned by the kubelet.  The playbook will also fail if the
filtering results in no hosts to prevent unknowingly upgrading all
nodes.