openshift-cluster/upgrades: Allow openshift_kubelet_name_override when filtering
When filtering the list of nodes to upgrade based on using the
openshift_upgrade_nodes_label variable, the node names may not match if
the kubelet returns the short node name.

 - `hostname` = fqdn          (infra-0.example.com)
 - `hostname -f` = fqdn       (infra-0.example.com)
 - `hostnamectl` = fqdn       (infra-0.example.com)
 - `oc get node` = short name (infra-0)

This changes the filter to test if l_kubelet_node_name is in the node
name returned by the kubelet.  l_kubelet_node_name can be overridden by
openshift_kubelet_name_override or it will default to
openshift.node.nodename provided by openshift_facts.

The playbook will also fail if the filtering results in no hosts to
prevent unknowingly upgrading all nodes.