Expose custom URL configuration for openshift-cluster-monitoring
The cluster-monitoring-operator supports configuring custom host names
for Prometheus, Alertmanager, and Grafana, but the respective
configuration options are not exposed by the
openshift-cluster-monitoring role.

This commit introduces variables
`openshift_cluster_monitoring_prometheus_external_url`,
`openshift_cluster_monitoring_alertmanager_external_url`, and
`openshift_cluster_monitoring_grafana_external_url`, which can be used
to configure non-standard hostnames for Prometheus, Alertmanager, and
Grafana respectively.

The cluster-monitoring-operator provides a field called `hostport` in each
service configuration, which is used to set the host on the route
objects for the services, and also the `externalUrl` field in the
Prometheus and Alertmanager custom resources.

To allow the cluster-monitoring-operator to configure a non-standard
host name on the routes, we need to grant it the RBAC permission
`routes/custom-host`.

This commit does this by adding the permission in the
cluster-monitoring-operator's ClusterRole.